Added all the files
The interaction is shooting, which is primarily intended to be used to destroy the enemy tanks. You press and hold the space bar, and the longer you hold it the farther the projectile goes. There is a secondary function of this interaction, as the player can destroy their own tank, which is an intended negative effect. The ai controlled enemy tanks can also shoot, potentially destroying themselves, other ai tanks, or the player tank, but there is some issue with their code so their projectile fires constantly and at one fixed range. Your attacks can destroy enemy tanks, as demonstrated by their health ring, or knock them around, which can be very helpful while being pursued. This could be extended out into destroying the buildings around the map.
Note that I gave up on audio, if you play it I recommend turning it off as I did in my video recording.
https://play.unity.com/en/games/b15255c5-970a-4b80-8962-3f7e5ef0b502/trayouras-tanks